backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
scikit-learn
spacy
numpy
